src/app/guard/auth.guard.ts
/***** Modules *****/
import { Injectable } from '@angular/core';
import { CanActivate, ActivatedRouteSnapshot, RouterStateSnapshot, UrlTree, Router } from '@angular/router';
import { Observable } from 'rxjs';

@Injectable({
  providedIn: 'root'
})
export class AuthGuard implements CanActivate {
  constructor(private router: Router) { }

  canActivate(
    next: ActivatedRouteSnapshot,
    state: RouterStateSnapshot): Observable<boolean | UrlTree> | Promise<boolean | UrlTree> | boolean | UrlTree {
    // navigate to farmers login page as user is not authenticated
    if(localStorage.getItem('isLoggedIn') === 'true'){
      return true;
    }
    this.router.navigate(['/']);
    return false;
  }

}
